Rename spec to model
