+++ src/controller_examples/rabbitmq_controller/exec/resource/erlang_cookie.rs
@@ -9,7 +9,7 @@ use crate::kubernetes_api_objects::{
     volume::*,
 };
 use crate::rabbitmq_controller::exec::resource::default_user_secret::DefaultUserSecretBuilder;
-use crate::rabbitmq_controller::spec::resource as spec_resource;
+use crate::rabbitmq_controller::model::resource as model_resource;
 use crate::rabbitmq_controller::trusted::exec_types::*;
 use crate::rabbitmq_controller::trusted::spec_types::RabbitmqClusterView;
 use crate::rabbitmq_controller::trusted::step::*;
@@ -24,7 +24,7 @@ verus! {
 
 pub struct ErlangCookieBuilder {}
 
-impl ResourceBuilder<RabbitmqCluster, RabbitmqReconcileState, spec_resource::ErlangCookieBuilder> for ErlangCookieBuilder {
+impl ResourceBuilder<RabbitmqCluster, RabbitmqReconcileState, model_resource::ErlangCookieBuilder> for ErlangCookieBuilder {
     open spec fn requirements(rabbitmq: RabbitmqClusterView) -> bool {
         &&& rabbitmq.metadata.name.is_Some()
         &&& rabbitmq.metadata.namespace.is_Some()
@@ -85,7 +85,7 @@ pub fn update_erlang_secret(rabbitmq: &RabbitmqCluster, found_erlang_secret: Sec
         rabbitmq@.metadata.name.is_Some(),
         rabbitmq@.metadata.namespace.is_Some(),
     ensures
-        secret@ == spec_resource::update_erlang_secret(rabbitmq@, found_erlang_secret@),
+        secret@ == model_resource::update_erlang_secret(rabbitmq@, found_erlang_secret@),
 {
     let mut erlang_secret = found_erlang_secret.clone();
     let made_secret = make_erlang_secret(rabbitmq);
@@ -96,7 +96,7 @@ pub fn update_erlang_secret(rabbitmq: &RabbitmqCluster, found_erlang_secret: Sec
         proof {
             assert_seqs_equal!(
                 owner_references@.map_values(|owner_ref: OwnerReference| owner_ref@),
-                spec_resource::update_erlang_secret(rabbitmq@, found_erlang_secret@).metadata.owner_references.get_Some_0()
+                model_resource::update_erlang_secret(rabbitmq@, found_erlang_secret@).metadata.owner_references.get_Some_0()
             );
         }
         metadata.set_owner_references(make_owner_references(rabbitmq));
@@ -113,7 +113,7 @@ pub fn make_erlang_secret_name(rabbitmq: &RabbitmqCluster) -> (name: String)
         rabbitmq@.metadata.name.is_Some(),
         rabbitmq@.metadata.namespace.is_Some(),
     ensures
-        name@ == spec_resource::make_erlang_secret_name(rabbitmq@),
+        name@ == model_resource::make_erlang_secret_name(rabbitmq@),
 {
     rabbitmq.metadata().name().unwrap().concat(new_strlit("-erlang-cookie"))
 }
@@ -123,7 +123,7 @@ pub fn make_erlang_secret(rabbitmq: &RabbitmqCluster) -> (secret: Secret)
         rabbitmq@.metadata.name.is_Some(),
         rabbitmq@.metadata.namespace.is_Some(),
     ensures
-        secret@ == spec_resource::make_erlang_secret(rabbitmq@)
+        secret@ == model_resource::make_erlang_secret(rabbitmq@)
 {
     let mut data = StringMap::empty();
     let cookie = random_encoded_string(24);
