+++ src/controller_examples/rabbitmq_controller/mod.rs
@@ -1,6 +1,6 @@
 // Copyright 2022 VMware, Inc.
 // SPDX-License-Identifier: MIT
 pub mod exec;
+pub mod model;
 pub mod proof;
-pub mod spec;
 pub mod trusted;
