+++ src/controller_examples/rabbitmq_controller/model/mod.rs
