+++ src/controller_examples/rabbitmq_controller/model/resource/common.rs
