+++ src/controller_examples/rabbitmq_controller/model/resource/erlang_cookie.rs
@@ -8,7 +8,7 @@ use crate::kubernetes_api_objects::{
     volume::*,
 };
 use crate::kubernetes_cluster::spec::message::*;
-use crate::rabbitmq_controller::spec::resource::config_map::ServerConfigMapBuilder;
+use crate::rabbitmq_controller::model::resource::default_user_secret::DefaultUserSecretBuilder;
 use crate::rabbitmq_controller::trusted::spec_types::*;
 use crate::rabbitmq_controller::trusted::step::*;
 use crate::reconciler::spec::{io::*, reconciler::*, resource_builder::*};
@@ -20,34 +20,34 @@ use vstd::string::*;
 
 verus! {
 
-pub struct PluginsConfigMapBuilder {}
+pub struct ErlangCookieBuilder {}
 
-impl ResourceBuilder<RabbitmqClusterView, RabbitmqReconcileState> for PluginsConfigMapBuilder {
+impl ResourceBuilder<RabbitmqClusterView, RabbitmqReconcileState> for ErlangCookieBuilder {
     open spec fn get_request(rabbitmq: RabbitmqClusterView) -> GetRequest {
-        GetRequest { key: make_plugins_config_map_key(rabbitmq) }
+        GetRequest { key: make_erlang_secret_key(rabbitmq) }
     }
 
     open spec fn make(rabbitmq: RabbitmqClusterView, state: RabbitmqReconcileState) -> Result<DynamicObjectView, ()> {
-        Ok(make_plugins_config_map(rabbitmq).marshal())
+        Ok(make_erlang_secret(rabbitmq).marshal())
     }
 
     open spec fn update(rabbitmq: RabbitmqClusterView, state: RabbitmqReconcileState, obj: DynamicObjectView) -> Result<DynamicObjectView, ()> {
-        let cm = ConfigMapView::unmarshal(obj);
-        if cm.is_Ok() {
-            Ok(update_plugins_config_map(rabbitmq, cm.get_Ok_0()).marshal())
+        let secret = SecretView::unmarshal(obj);
+        if secret.is_Ok() {
+            Ok(update_erlang_secret(rabbitmq, secret.get_Ok_0()).marshal())
         } else {
             Err(())
         }
     }
 
     open spec fn state_after_create(rabbitmq: RabbitmqClusterView, obj: DynamicObjectView, state: RabbitmqReconcileState) -> (res: Result<(RabbitmqReconcileState, Option<APIRequest>), ()>) {
-        let cm = ConfigMapView::unmarshal(obj);
-        if cm.is_Ok() {
+        let secret = SecretView::unmarshal(obj);
+        if secret.is_Ok() {
             let state_prime = RabbitmqReconcileState {
-                reconcile_step: RabbitmqReconcileStep::AfterKRequestStep(ActionKind::Get, SubResource::ServerConfigMap),
+                reconcile_step: RabbitmqReconcileStep::AfterKRequestStep(ActionKind::Get, SubResource::DefaultUserSecret),
                 ..state
             };
-            let req = APIRequest::GetRequest(ServerConfigMapBuilder::get_request(rabbitmq));
+            let req = APIRequest::GetRequest(DefaultUserSecretBuilder::get_request(rabbitmq));
             Ok((state_prime, Some(req)))
         } else {
             Err(())
@@ -55,13 +55,13 @@ impl ResourceBuilder<RabbitmqClusterView, RabbitmqReconcileState> for PluginsCon
     }
 
     open spec fn state_after_update(rabbitmq: RabbitmqClusterView, obj: DynamicObjectView, state: RabbitmqReconcileState) -> (res: Result<(RabbitmqReconcileState, Option<APIRequest>), ()>) {
-        let cm = ConfigMapView::unmarshal(obj);
-        if cm.is_Ok() {
+        let secret = SecretView::unmarshal(obj);
+        if secret.is_Ok() {
             let state_prime = RabbitmqReconcileState {
-                reconcile_step: RabbitmqReconcileStep::AfterKRequestStep(ActionKind::Get, SubResource::ServerConfigMap),
+                reconcile_step: RabbitmqReconcileStep::AfterKRequestStep(ActionKind::Get, SubResource::DefaultUserSecret),
                 ..state
             };
-            let req = APIRequest::GetRequest(ServerConfigMapBuilder::get_request(rabbitmq));
+            let req = APIRequest::GetRequest(DefaultUserSecretBuilder::get_request(rabbitmq));
             Ok((state_prime, Some(req)))
         } else {
             Err(())
@@ -69,60 +69,52 @@ impl ResourceBuilder<RabbitmqClusterView, RabbitmqReconcileState> for PluginsCon
     }
 }
 
-pub open spec fn make_plugins_config_map_name(rabbitmq: RabbitmqClusterView) -> StringView
+pub open spec fn make_erlang_secret_name(rabbitmq: RabbitmqClusterView) -> StringView
     recommends
         rabbitmq.metadata.name.is_Some(),
 {
-    rabbitmq.metadata.name.get_Some_0() + new_strlit("-plugins-conf")@
+    rabbitmq.metadata.name.get_Some_0() + new_strlit("-erlang-cookie")@
 }
 
-pub open spec fn make_plugins_config_map_key(rabbitmq: RabbitmqClusterView) -> ObjectRef
+pub open spec fn make_erlang_secret_key(rabbitmq: RabbitmqClusterView) -> ObjectRef
     recommends
         rabbitmq.metadata.name.is_Some(),
         rabbitmq.metadata.namespace.is_Some(),
 {
     ObjectRef {
-        kind: ConfigMapView::kind(),
-        name: make_plugins_config_map_name(rabbitmq),
+        kind: SecretView::kind(),
+        name: make_erlang_secret_name(rabbitmq),
         namespace: rabbitmq.metadata.namespace.get_Some_0(),
     }
 }
 
-pub open spec fn update_plugins_config_map(rabbitmq: RabbitmqClusterView, found_config_map: ConfigMapView) -> ConfigMapView
+pub open spec fn update_erlang_secret(rabbitmq: RabbitmqClusterView, found_erlang_secret: SecretView) -> SecretView
     recommends
         rabbitmq.metadata.name.is_Some(),
         rabbitmq.metadata.namespace.is_Some(),
 {
-    let made_config_map = make_plugins_config_map(rabbitmq);
-    ConfigMapView {
-        data: made_config_map.data,
+    let made_erlang_secret = make_erlang_secret(rabbitmq);
+    SecretView {
         metadata: ObjectMetaView {
             owner_references: Some(make_owner_references(rabbitmq)),
             finalizers: None,
-            labels: made_config_map.metadata.labels,
-            annotations: made_config_map.metadata.annotations,
-            ..found_config_map.metadata
+            labels: made_erlang_secret.metadata.labels,
+            annotations: made_erlang_secret.metadata.annotations,
+            ..found_erlang_secret.metadata
         },
-        ..found_config_map
+        ..found_erlang_secret
     }
 }
 
-pub open spec fn make_plugins_config_map(rabbitmq: RabbitmqClusterView) -> ConfigMapView
+pub open spec fn make_erlang_secret(rabbitmq: RabbitmqClusterView) -> SecretView
     recommends
         rabbitmq.metadata.name.is_Some(),
         rabbitmq.metadata.namespace.is_Some(),
 {
-    ConfigMapView::default()
-        .set_metadata(ObjectMetaView::default()
-            .set_name(make_plugins_config_map_name(rabbitmq))
-            .set_namespace(rabbitmq.metadata.namespace.get_Some_0())
-            .set_owner_references(make_owner_references(rabbitmq))
-            .set_labels(make_labels(rabbitmq))
-            .set_annotations(rabbitmq.spec.annotations)
-        )
-        .set_data(Map::empty()
-            .insert(new_strlit("enabled_plugins")@, new_strlit("[rabbitmq_peer_discovery_k8s,rabbitmq_prometheus,rabbitmq_management].")@)
-        )
+    let cookie = random_encoded_string(24);
+    let data = Map::empty()
+        .insert(new_strlit(".erlang.cookie")@, cookie);
+    make_secret(rabbitmq, make_erlang_secret_name(rabbitmq), data)
 }
 
 }
