+++ src/controller_examples/rabbitmq_controller/model/resource/headless_service.rs
@@ -8,7 +8,7 @@ use crate::kubernetes_api_objects::{
     volume::*,
 };
 use crate::kubernetes_cluster::spec::message::*;
-use crate::rabbitmq_controller::spec::resource::erlang_cookie::ErlangCookieBuilder;
+use crate::rabbitmq_controller::model::resource::service::ServiceBuilder;
 use crate::rabbitmq_controller::trusted::spec_types::*;
 use crate::rabbitmq_controller::trusted::step::*;
 use crate::reconciler::spec::{io::*, reconciler::*, resource_builder::*};
@@ -20,22 +20,22 @@ use vstd::string::*;
 
 verus! {
 
-pub struct ServiceBuilder {}
+pub struct HeadlessServiceBuilder {}
 
-impl ResourceBuilder<RabbitmqClusterView, RabbitmqReconcileState> for ServiceBuilder {
+impl ResourceBuilder<RabbitmqClusterView, RabbitmqReconcileState> for HeadlessServiceBuilder {
     open spec fn get_request(rabbitmq: RabbitmqClusterView) -> GetRequest {
-        GetRequest { key: make_main_service_key(rabbitmq) }
+        GetRequest { key: make_headless_service_key(rabbitmq) }
     }
 
     open spec fn make(rabbitmq: RabbitmqClusterView, state: RabbitmqReconcileState) -> Result<DynamicObjectView, ()> {
-        Ok(make_main_service(rabbitmq).marshal())
+        Ok(make_headless_service(rabbitmq).marshal())
     }
 
     open spec fn update(rabbitmq: RabbitmqClusterView, state: RabbitmqReconcileState, obj: DynamicObjectView) -> Result<DynamicObjectView, ()> {
         let service = ServiceView::unmarshal(obj);
         let found_service = service.get_Ok_0();
         if service.is_Ok() && found_service.spec.is_Some() {
-            Ok(update_main_service(rabbitmq, found_service).marshal())
+            Ok(update_headless_service(rabbitmq, found_service).marshal())
         } else {
             Err(())
         }
@@ -45,10 +45,10 @@ impl ResourceBuilder<RabbitmqClusterView, RabbitmqReconcileState> for ServiceBui
         let service = ServiceView::unmarshal(obj);
         if service.is_Ok() {
             let state_prime = RabbitmqReconcileState {
-                reconcile_step: RabbitmqReconcileStep::AfterKRequestStep(ActionKind::Get, SubResource::ErlangCookieSecret),
+                reconcile_step: RabbitmqReconcileStep::AfterKRequestStep(ActionKind::Get, SubResource::Service),
                 ..state
             };
-            let req = APIRequest::GetRequest(ErlangCookieBuilder::get_request(rabbitmq));
+            let req = APIRequest::GetRequest(ServiceBuilder::get_request(rabbitmq));
             Ok((state_prime, Some(req)))
         } else {
             Err(())
@@ -59,10 +59,10 @@ impl ResourceBuilder<RabbitmqClusterView, RabbitmqReconcileState> for ServiceBui
         let service = ServiceView::unmarshal(obj);
         if service.is_Ok() {
             let state_prime = RabbitmqReconcileState {
-                reconcile_step: RabbitmqReconcileStep::AfterKRequestStep(ActionKind::Get, SubResource::ErlangCookieSecret),
+                reconcile_step: RabbitmqReconcileStep::AfterKRequestStep(ActionKind::Get, SubResource::Service),
                 ..state
             };
-            let req = APIRequest::GetRequest(ErlangCookieBuilder::get_request(rabbitmq));
+            let req = APIRequest::GetRequest(ServiceBuilder::get_request(rabbitmq));
             Ok((state_prime, Some(req)))
         } else {
             Err(())
@@ -70,61 +70,60 @@ impl ResourceBuilder<RabbitmqClusterView, RabbitmqReconcileState> for ServiceBui
     }
 }
 
-pub open spec fn make_main_service_name(rabbitmq: RabbitmqClusterView) -> StringView
+pub open spec fn make_headless_service_name(rabbitmq: RabbitmqClusterView) -> StringView
     recommends
         rabbitmq.metadata.name.is_Some(),
 {
-    rabbitmq.metadata.name.get_Some_0() + new_strlit("-client")@
+    rabbitmq.metadata.name.get_Some_0() + new_strlit("-nodes")@
 }
 
-pub open spec fn make_main_service_key(rabbitmq: RabbitmqClusterView) -> ObjectRef
+pub open spec fn make_headless_service_key(rabbitmq: RabbitmqClusterView) -> ObjectRef
     recommends
         rabbitmq.metadata.name.is_Some(),
         rabbitmq.metadata.namespace.is_Some(),
 {
     ObjectRef {
         kind: ServiceView::kind(),
-        name: make_main_service_name(rabbitmq),
+        name: make_headless_service_name(rabbitmq),
         namespace: rabbitmq.metadata.namespace.get_Some_0(),
     }
 }
 
-pub open spec fn update_main_service(rabbitmq: RabbitmqClusterView, found_main_service: ServiceView) -> ServiceView
+pub open spec fn update_headless_service(rabbitmq: RabbitmqClusterView, found_headless_service: ServiceView) -> ServiceView
     recommends
         rabbitmq.metadata.name.is_Some(),
         rabbitmq.metadata.namespace.is_Some(),
-        found_main_service.spec.is_Some(),
+        found_headless_service.spec.is_Some(),
 {
-    let made_main_service = make_main_service(rabbitmq);
+    let made_headless_service = make_headless_service(rabbitmq);
     ServiceView {
         metadata: ObjectMetaView {
             owner_references: Some(make_owner_references(rabbitmq)),
             finalizers: None,
-            labels: made_main_service.metadata.labels,
-            annotations: made_main_service.metadata.annotations,
-            ..found_main_service.metadata
+            labels: made_headless_service.metadata.labels,
+            annotations: made_headless_service.metadata.annotations,
+            ..found_headless_service.metadata
         },
         spec: Some(ServiceSpecView {
-            ports: made_main_service.spec.get_Some_0().ports,
-            selector: made_main_service.spec.get_Some_0().selector,
-            publish_not_ready_addresses: made_main_service.spec.get_Some_0().publish_not_ready_addresses,
-            ..found_main_service.spec.get_Some_0()
+            ports: made_headless_service.spec.get_Some_0().ports,
+            selector: made_headless_service.spec.get_Some_0().selector,
+            publish_not_ready_addresses: made_headless_service.spec.get_Some_0().publish_not_ready_addresses,
+            ..found_headless_service.spec.get_Some_0()
         }),
-        ..found_main_service
+        ..found_headless_service
     }
 }
 
-pub open spec fn make_main_service(rabbitmq: RabbitmqClusterView) -> ServiceView
+pub open spec fn make_headless_service(rabbitmq: RabbitmqClusterView) -> ServiceView
     recommends
         rabbitmq.metadata.name.is_Some(),
         rabbitmq.metadata.namespace.is_Some(),
 {
-    let ports = seq![
-        ServicePortView::default().set_name(new_strlit("amqp")@).set_port(5672).set_app_protocol(new_strlit("amqp")@),
-        ServicePortView::default().set_name(new_strlit("management")@).set_port(15672).set_app_protocol(new_strlit("http")@),
-        ServicePortView::default().set_name(new_strlit("prometheus")@).set_port(15692).set_app_protocol(new_strlit("prometheus.io/metrics")@),
+    let mut ports = seq![
+        ServicePortView::default().set_name(new_strlit("epmd")@).set_port(4369),
+        ServicePortView::default().set_name(new_strlit("cluster-rpc")@).set_port(25672)
     ];
-    make_service(rabbitmq, make_main_service_name(rabbitmq), ports, true)
+    make_service(rabbitmq, make_headless_service_name(rabbitmq), ports, false)
 }
 
 }
