+++ src/controller_examples/rabbitmq_controller/model/resource/rabbitmq_plugins.rs
@@ -8,7 +8,7 @@ use crate::kubernetes_api_objects::{
     volume::*,
 };
 use crate::kubernetes_cluster::spec::message::*;
-use crate::rabbitmq_controller::spec::resource::StatefulSetBuilder;
+use crate::rabbitmq_controller::model::resource::config_map::ServerConfigMapBuilder;
 use crate::rabbitmq_controller::trusted::spec_types::*;
 use crate::rabbitmq_controller::trusted::step::*;
 use crate::reconciler::spec::{io::*, reconciler::*, resource_builder::*};
@@ -20,34 +20,34 @@ use vstd::string::*;
 
 verus! {
 
-pub struct RoleBindingBuilder {}
+pub struct PluginsConfigMapBuilder {}
 
-impl ResourceBuilder<RabbitmqClusterView, RabbitmqReconcileState> for RoleBindingBuilder {
+impl ResourceBuilder<RabbitmqClusterView, RabbitmqReconcileState> for PluginsConfigMapBuilder {
     open spec fn get_request(rabbitmq: RabbitmqClusterView) -> GetRequest {
-        GetRequest { key: make_role_binding_key(rabbitmq) }
+        GetRequest { key: make_plugins_config_map_key(rabbitmq) }
     }
 
     open spec fn make(rabbitmq: RabbitmqClusterView, state: RabbitmqReconcileState) -> Result<DynamicObjectView, ()> {
-        Ok(make_role_binding(rabbitmq).marshal())
+        Ok(make_plugins_config_map(rabbitmq).marshal())
     }
 
     open spec fn update(rabbitmq: RabbitmqClusterView, state: RabbitmqReconcileState, obj: DynamicObjectView) -> Result<DynamicObjectView, ()> {
-        let rb = RoleBindingView::unmarshal(obj);
-        if rb.is_Ok() {
-            Ok(update_role_binding(rabbitmq, rb.get_Ok_0()).marshal())
+        let cm = ConfigMapView::unmarshal(obj);
+        if cm.is_Ok() {
+            Ok(update_plugins_config_map(rabbitmq, cm.get_Ok_0()).marshal())
         } else {
             Err(())
         }
     }
 
     open spec fn state_after_create(rabbitmq: RabbitmqClusterView, obj: DynamicObjectView, state: RabbitmqReconcileState) -> (res: Result<(RabbitmqReconcileState, Option<APIRequest>), ()>) {
-        let rb = RoleBindingView::unmarshal(obj);
-        if rb.is_Ok() {
+        let cm = ConfigMapView::unmarshal(obj);
+        if cm.is_Ok() {
             let state_prime = RabbitmqReconcileState {
-                reconcile_step: RabbitmqReconcileStep::AfterKRequestStep(ActionKind::Get, SubResource::StatefulSet),
+                reconcile_step: RabbitmqReconcileStep::AfterKRequestStep(ActionKind::Get, SubResource::ServerConfigMap),
                 ..state
             };
-            let req = APIRequest::GetRequest(StatefulSetBuilder::get_request(rabbitmq));
+            let req = APIRequest::GetRequest(ServerConfigMapBuilder::get_request(rabbitmq));
             Ok((state_prime, Some(req)))
         } else {
             Err(())
@@ -55,13 +55,13 @@ impl ResourceBuilder<RabbitmqClusterView, RabbitmqReconcileState> for RoleBindin
     }
 
     open spec fn state_after_update(rabbitmq: RabbitmqClusterView, obj: DynamicObjectView, state: RabbitmqReconcileState) -> (res: Result<(RabbitmqReconcileState, Option<APIRequest>), ()>) {
-        let rb = RoleBindingView::unmarshal(obj);
-        if rb.is_Ok() {
+        let cm = ConfigMapView::unmarshal(obj);
+        if cm.is_Ok() {
             let state_prime = RabbitmqReconcileState {
-                reconcile_step: RabbitmqReconcileStep::AfterKRequestStep(ActionKind::Get, SubResource::StatefulSet),
+                reconcile_step: RabbitmqReconcileStep::AfterKRequestStep(ActionKind::Get, SubResource::ServerConfigMap),
                 ..state
             };
-            let req = APIRequest::GetRequest(StatefulSetBuilder::get_request(rabbitmq));
+            let req = APIRequest::GetRequest(ServerConfigMapBuilder::get_request(rabbitmq));
             Ok((state_prime, Some(req)))
         } else {
             Err(())
@@ -69,65 +69,60 @@ impl ResourceBuilder<RabbitmqClusterView, RabbitmqReconcileState> for RoleBindin
     }
 }
 
-pub open spec fn make_role_binding_name(rabbitmq: RabbitmqClusterView) -> StringView
+pub open spec fn make_plugins_config_map_name(rabbitmq: RabbitmqClusterView) -> StringView
     recommends
         rabbitmq.metadata.name.is_Some(),
 {
-    rabbitmq.metadata.name.get_Some_0() + new_strlit("-server")@
+    rabbitmq.metadata.name.get_Some_0() + new_strlit("-plugins-conf")@
 }
 
-pub open spec fn make_role_binding_key(rabbitmq: RabbitmqClusterView) -> ObjectRef
+pub open spec fn make_plugins_config_map_key(rabbitmq: RabbitmqClusterView) -> ObjectRef
     recommends
         rabbitmq.metadata.name.is_Some(),
         rabbitmq.metadata.namespace.is_Some(),
 {
     ObjectRef {
-        kind: RoleBindingView::kind(),
-        name: make_role_binding_name(rabbitmq),
+        kind: ConfigMapView::kind(),
+        name: make_plugins_config_map_name(rabbitmq),
         namespace: rabbitmq.metadata.namespace.get_Some_0(),
     }
 }
 
-pub open spec fn update_role_binding(rabbitmq: RabbitmqClusterView, found_role_binding: RoleBindingView) -> RoleBindingView
+pub open spec fn update_plugins_config_map(rabbitmq: RabbitmqClusterView, found_config_map: ConfigMapView) -> ConfigMapView
     recommends
         rabbitmq.metadata.name.is_Some(),
         rabbitmq.metadata.namespace.is_Some(),
 {
-    let made_role_binding = make_role_binding(rabbitmq);
-    RoleBindingView {
+    let made_config_map = make_plugins_config_map(rabbitmq);
+    ConfigMapView {
+        data: made_config_map.data,
         metadata: ObjectMetaView {
             owner_references: Some(make_owner_references(rabbitmq)),
             finalizers: None,
-            labels: made_role_binding.metadata.labels,
-            annotations: made_role_binding.metadata.annotations,
-            ..found_role_binding.metadata
+            labels: made_config_map.metadata.labels,
+            annotations: made_config_map.metadata.annotations,
+            ..found_config_map.metadata
         },
-        subjects: made_role_binding.subjects,
-        ..found_role_binding
+        ..found_config_map
     }
 }
 
-pub open spec fn make_role_binding(rabbitmq: RabbitmqClusterView) -> RoleBindingView
+pub open spec fn make_plugins_config_map(rabbitmq: RabbitmqClusterView) -> ConfigMapView
     recommends
         rabbitmq.metadata.name.is_Some(),
         rabbitmq.metadata.namespace.is_Some(),
 {
-    RoleBindingView::default()
+    ConfigMapView::default()
         .set_metadata(ObjectMetaView::default()
-            .set_name(make_role_binding_name(rabbitmq))
+            .set_name(make_plugins_config_map_name(rabbitmq))
             .set_namespace(rabbitmq.metadata.namespace.get_Some_0())
             .set_owner_references(make_owner_references(rabbitmq))
             .set_labels(make_labels(rabbitmq))
             .set_annotations(rabbitmq.spec.annotations)
-        ).set_role_ref(RoleRefView::default()
-            .set_api_group(new_strlit("rbac.authorization.k8s.io")@)
-            .set_kind(new_strlit("Role")@)
-            .set_name(rabbitmq.metadata.name.get_Some_0() + new_strlit("-peer-discovery")@)
-        ).set_subjects(seq![SubjectView::default()
-            .set_kind(new_strlit("ServiceAccount")@)
-            .set_name(rabbitmq.metadata.name.get_Some_0() + new_strlit("-server")@)
-            .set_namespace(rabbitmq.metadata.namespace.get_Some_0())
-        ])
+        )
+        .set_data(Map::empty()
+            .insert(new_strlit("enabled_plugins")@, new_strlit("[rabbitmq_peer_discovery_k8s,rabbitmq_prometheus,rabbitmq_management].")@)
+        )
 }
 
 }
