+++ src/controller_examples/rabbitmq_controller/model/resource/service_account.rs
@@ -8,7 +8,7 @@ use crate::kubernetes_api_objects::{
     volume::*,
 };
 use crate::kubernetes_cluster::spec::message::*;
-use crate::rabbitmq_controller::spec::resource::role_binding::RoleBindingBuilder;
+use crate::rabbitmq_controller::model::resource::role::RoleBuilder;
 use crate::rabbitmq_controller::trusted::spec_types::*;
 use crate::rabbitmq_controller::trusted::step::*;
 use crate::reconciler::spec::{io::*, reconciler::*, resource_builder::*};
@@ -20,34 +20,34 @@ use vstd::string::*;
 
 verus! {
 
-pub struct RoleBuilder {}
+pub struct ServiceAccountBuilder {}
 
-impl ResourceBuilder<RabbitmqClusterView, RabbitmqReconcileState> for RoleBuilder {
+impl ResourceBuilder<RabbitmqClusterView, RabbitmqReconcileState> for ServiceAccountBuilder {
     open spec fn get_request(rabbitmq: RabbitmqClusterView) -> GetRequest {
-        GetRequest { key: make_role_key(rabbitmq) }
+        GetRequest { key: make_service_account_key(rabbitmq) }
     }
 
     open spec fn make(rabbitmq: RabbitmqClusterView, state: RabbitmqReconcileState) -> Result<DynamicObjectView, ()> {
-        Ok(make_role(rabbitmq).marshal())
+        Ok(make_service_account(rabbitmq).marshal())
     }
 
     open spec fn update(rabbitmq: RabbitmqClusterView, state: RabbitmqReconcileState, obj: DynamicObjectView) -> Result<DynamicObjectView, ()> {
-        let role = RoleView::unmarshal(obj);
-        if role.is_Ok() {
-            Ok(update_role(rabbitmq, role.get_Ok_0()).marshal())
+        let sa = ServiceAccountView::unmarshal(obj);
+        if sa.is_Ok() {
+            Ok(update_service_account(rabbitmq, sa.get_Ok_0()).marshal())
         } else {
             Err(())
         }
     }
 
     open spec fn state_after_create(rabbitmq: RabbitmqClusterView, obj: DynamicObjectView, state: RabbitmqReconcileState) -> (res: Result<(RabbitmqReconcileState, Option<APIRequest>), ()>) {
-        let role = RoleView::unmarshal(obj);
-        if role.is_Ok() {
+        let sa = ServiceAccountView::unmarshal(obj);
+        if sa.is_Ok() {
             let state_prime = RabbitmqReconcileState {
-                reconcile_step: RabbitmqReconcileStep::AfterKRequestStep(ActionKind::Get, SubResource::RoleBinding),
+                reconcile_step: RabbitmqReconcileStep::AfterKRequestStep(ActionKind::Get, SubResource::Role),
                 ..state
             };
-            let req = APIRequest::GetRequest(RoleBindingBuilder::get_request(rabbitmq));
+            let req = APIRequest::GetRequest(RoleBuilder::get_request(rabbitmq));
             Ok((state_prime, Some(req)))
         } else {
             Err(())
@@ -55,13 +55,13 @@ impl ResourceBuilder<RabbitmqClusterView, RabbitmqReconcileState> for RoleBuilde
     }
 
     open spec fn state_after_update(rabbitmq: RabbitmqClusterView, obj: DynamicObjectView, state: RabbitmqReconcileState) -> (res: Result<(RabbitmqReconcileState, Option<APIRequest>), ()>) {
-        let role = RoleView::unmarshal(obj);
-        if role.is_Ok() {
+        let sa = ServiceAccountView::unmarshal(obj);
+        if sa.is_Ok() {
             let state_prime = RabbitmqReconcileState {
-                reconcile_step: RabbitmqReconcileStep::AfterKRequestStep(ActionKind::Get, SubResource::RoleBinding),
+                reconcile_step: RabbitmqReconcileStep::AfterKRequestStep(ActionKind::Get, SubResource::Role),
                 ..state
             };
-            let req = APIRequest::GetRequest(RoleBindingBuilder::get_request(rabbitmq));
+            let req = APIRequest::GetRequest(RoleBuilder::get_request(rabbitmq));
             Ok((state_prime, Some(req)))
         } else {
             Err(())
@@ -69,62 +69,59 @@ impl ResourceBuilder<RabbitmqClusterView, RabbitmqReconcileState> for RoleBuilde
     }
 }
 
-pub open spec fn make_role_name(rabbitmq: RabbitmqClusterView) -> StringView
+pub open spec fn make_service_account_name(rabbitmq: RabbitmqClusterView) -> StringView
     recommends
         rabbitmq.metadata.name.is_Some(),
 {
-    rabbitmq.metadata.name.get_Some_0() + new_strlit("-peer-discovery")@
+    rabbitmq.metadata.name.get_Some_0() + new_strlit("-server")@
 }
 
-pub open spec fn make_role_key(rabbitmq: RabbitmqClusterView) -> ObjectRef
+pub open spec fn make_service_account_key(rabbitmq: RabbitmqClusterView) -> ObjectRef
     recommends
         rabbitmq.metadata.name.is_Some(),
         rabbitmq.metadata.namespace.is_Some(),
 {
     ObjectRef {
-        kind: RoleView::kind(),
-        name: make_role_name(rabbitmq),
+        kind: ServiceAccountView::kind(),
+        name: make_service_account_name(rabbitmq),
         namespace: rabbitmq.metadata.namespace.get_Some_0(),
     }
 }
 
-pub open spec fn update_role(rabbitmq: RabbitmqClusterView, found_role: RoleView) -> RoleView
+pub open spec fn update_service_account(rabbitmq: RabbitmqClusterView, found_service_account: ServiceAccountView) -> ServiceAccountView
     recommends
         rabbitmq.metadata.name.is_Some(),
         rabbitmq.metadata.namespace.is_Some(),
 {
-    let made_role = make_role(rabbitmq);
-    RoleView {
-        policy_rules: made_role.policy_rules,
+    let made_service_account = make_service_account(rabbitmq);
+    ServiceAccountView {
         metadata: ObjectMetaView {
             owner_references: Some(make_owner_references(rabbitmq)),
             finalizers: None,
-            labels: made_role.metadata.labels,
-            annotations: made_role.metadata.annotations,
-            ..found_role.metadata
+            labels: made_service_account.metadata.labels,
+            annotations: made_service_account.metadata.annotations,
+            ..found_service_account.metadata
         },
-        ..found_role
+        ..found_service_account
     }
 }
 
-pub open spec fn make_role(rabbitmq: RabbitmqClusterView) -> RoleView
+pub open spec fn make_service_account(rabbitmq: RabbitmqClusterView) -> ServiceAccountView
     recommends
         rabbitmq.metadata.name.is_Some(),
         rabbitmq.metadata.namespace.is_Some(),
 {
-    RoleView::default()
-        .set_metadata(ObjectMetaView::default()
-            .set_name(make_role_name(rabbitmq))
-            .set_namespace(rabbitmq.metadata.namespace.get_Some_0())
-            .set_owner_references(make_owner_references(rabbitmq))
-            .set_labels(make_labels(rabbitmq))
-            .set_annotations(rabbitmq.spec.annotations)
-        ).set_policy_rules(
-            seq![
-                PolicyRuleView::default().set_api_groups(seq![new_strlit("")@]).set_resources(seq![new_strlit("endpoints")@]).set_verbs(seq![new_strlit("get")@]),
-                PolicyRuleView::default().set_api_groups(seq![new_strlit("")@]).set_resources(seq![new_strlit("events")@]).set_verbs(seq![new_strlit("create")@]),
-            ]
-        )
+    ServiceAccountView {
+        metadata: ObjectMetaView {
+            name: Some(make_service_account_name(rabbitmq)),
+            namespace: rabbitmq.metadata.namespace,
+            owner_references: Some(make_owner_references(rabbitmq)),
+            labels: Some(make_labels(rabbitmq)),
+            annotations: Some(rabbitmq.spec.annotations),
+            ..ObjectMetaView::default()
+        },
+        ..ServiceAccountView::default()
+    }
 }
 
 }
