+++ src/controller_examples/rabbitmq_controller/proof/helper_invariants/owner_ref.rs
@@ -14,12 +14,12 @@ use crate::kubernetes_cluster::spec::{
     message::*,
 };
 use crate::rabbitmq_controller::{
+    model::reconciler::*,
     proof::{
         helper_invariants::{predicate::*, proof::*},
         predicate::*,
         resource::*,
     },
-    spec::reconciler::*,
     trusted::{spec_types::*, step::*},
 };
 use crate::temporal_logic::{defs::*, rules::*};
