+++ src/controller_examples/rabbitmq_controller/proof/helper_invariants/unchangeable.rs
@@ -12,12 +12,12 @@ use crate::kubernetes_cluster::spec::{
     message::*,
 };
 use crate::rabbitmq_controller::{
+    model::{reconciler::*, resource::*},
     proof::{
         helper_invariants::{owner_ref::*, predicate::*, proof::*, validation::*},
         predicate::*,
         resource::*,
     },
-    spec::{reconciler::*, resource::*},
     trusted::{spec_types::*, step::*},
 };
 use crate::reconciler::spec::{reconciler::*, resource_builder::*};
