+++ src/controller_examples/rabbitmq_controller/proof/liveness/spec.rs
@@ -14,8 +14,8 @@ use crate::kubernetes_cluster::spec::{
     message::*,
 };
 use crate::rabbitmq_controller::{
+    model::reconciler::*,
     proof::{helper_invariants, liveness::terminate, predicate::*, resource::*},
-    spec::reconciler::*,
     trusted::{liveness_theorem::*, spec_types::*, step::*},
 };
 use crate::temporal_logic::{defs::*, rules::*};
