+++ src/controller_examples/rabbitmq_controller/proof/liveness/stateful_set_match.rs
@@ -13,13 +13,13 @@ use crate::kubernetes_cluster::spec::{
     message::*,
 };
 use crate::rabbitmq_controller::{
+    model::{reconciler::*, resource::*},
     proof::{
         helper_invariants,
         liveness::{resource_match::*, spec::assumption_and_invariants_of_all_phases},
         predicate::*,
         resource::*,
     },
-    spec::{reconciler::*, resource::*},
     trusted::{liveness_theorem::*, spec_types::*, step::*},
 };
 use crate::temporal_logic::{defs::*, rules::*};
