+++ src/controller_examples/rabbitmq_controller/proof/liveness/terminate.rs
@@ -11,8 +11,8 @@ use crate::kubernetes_cluster::spec::{
     message::*,
 };
 use crate::rabbitmq_controller::{
+    model::reconciler::*,
     proof::predicate::*,
-    spec::reconciler::*,
     trusted::{spec_types::*, step::*},
 };
 use crate::reconciler::spec::reconciler::*;
