+++ src/controller_examples/rabbitmq_controller/trusted/exec_types.rs
@@ -5,8 +5,7 @@ use crate::kubernetes_api_objects::{
     affinity::*, api_resource::*, common::*, dynamic::*, marshal::*, object_meta::*,
     owner_reference::*, resource::*, resource_requirements::*, stateful_set::*, toleration::*,
 };
-use crate::rabbitmq_controller::trusted::spec_types;
-use crate::rabbitmq_controller::trusted::step::*;
+use crate::rabbitmq_controller::trusted::{spec_types, step::*};
 use crate::vstd_ext::{string_map::*, string_view::*};
 use deps_hack::kube::Resource;
 use vstd::prelude::*;
